api/serializers.py
from rest_framework import serializers

class UserCredentialsSerializer(serializers.Serializer):
    email = serializers.CharField(max_length=100)
    password = serializers.CharField(max_length=100)
