requirements.txt
seedrcc
requests
beautifulsoup4
pytube
django
django-mathfilters
selenium
djangorestframework